% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrheatmap.R
\name{corrplot}
\alias{corrplot}
\title{Calculate and plot the correlation of a given dataset}
\usage{
corrplot(df, corr = "spearman", xlevels = NULL, ylevels = NULL)
}
\arguments{
\item{df}{Dataframe containing HI titer values, each column is a different strain}

\item{corr}{Correlation method "spearman" "pearson" "kendall", spearman is default}

\item{xlevels}{Can order or limit the levels on x axis (not done yet)}

\item{ylevels}{Can order or limit the levels on y axis (not done yet)}
}
\value{
The ggplot of the correlation heatmap
}
\description{
Calculate and plot the correlation of a given dataset
}
