% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_vipr_ncbi.R
\name{fncols}
\alias{fncols}
\title{Helps harmonize or expand column names, preparing for a summarize merge later}
\usage{
fncols(data, cname)
}
\arguments{
\item{data}{Data frame}

\item{cname}{Vector of strings of column names that need to exist in data frame}
}
\value{
New data frame containing a merge of original columns and cname
}
\description{
Helps harmonize or expand column names, preparing for a summarize merge later
}
