% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_vipr_ncbi.R
\name{format_ncbi_date}
\alias{format_ncbi_date}
\title{Format NCBI dates (YYYY-MM-DD) into some form of YYYY-MM-DD, with unknown set to XX}
\usage{
format_ncbi_date(vc)
}
\arguments{
\item{vc}{Character vector containing the ncbi date such as "2022-01-30" or "2022-01"}
}
\value{
The formatted date in "YYYY-MM-DD" format
}
\description{
Format NCBI dates (YYYY-MM-DD) into some form of YYYY-MM-DD, with unknown set to XX
}
