% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_vipr_ncbi.R
\name{format_vipr_date}
\alias{format_vipr_date}
\title{Format ViPR dates (MM/DD/YYYY) into some form of YYYY-MM-DD, with unknown set to XX}
\usage{
format_vipr_date(vc, delim = "/")
}
\arguments{
\item{vc}{Character vector containing the vipr date such as "01/30/2022" or "01/2022"}

\item{delim}{Delimiter of date values, which is "/" by default but could also be "-"}
}
\value{
The formatted date in "YYYY-MM-DD" format
}
\description{
Format ViPR dates (MM/DD/YYYY) into some form of YYYY-MM-DD, with unknown set to XX
}
