% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrheatmap.R
\name{HI_to_GMT}
\alias{HI_to_GMT}
\title{Converts HI Titer values to generalized mean titer (GMT)}
\usage{
HI_to_GMT(hi_titers)
}
\arguments{
\item{hi_titers}{HI Titer values, assumes all titers are above 0}
}
\value{
The GMT titer = log(HI titer/ 10, 2)
}
\description{
Converts HI Titer values to generalized mean titer (GMT)
}
