% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_vipr_ncbi.R
\name{merge_two}
\alias{merge_two}
\title{Merge two dataframes}
\usage{
merge_two(one_df, two_df)
}
\arguments{
\item{one_df}{Vector of values for a particular column}

\item{two_df}{Delimiter between unique values, will split original string into delims, and smash them together again}
}
\value{
Merged data frame
}
\description{
Merge two dataframes
}
