% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_vipr_ncbi.R
\name{read_delim_file}
\alias{read_delim_file}
\title{Read in delimited data
Drop any empty columns
Remove spaces from column names}
\usage{
read_delim_file(filename, delim = "\\t", type = "basic")
}
\arguments{
\item{filename}{Character vector containing path to file}

\item{delim}{Delimiter of file, tab by default}

\item{type}{Basic is a basic delimited file, but special handling for ncbi or vipr}
}
\value{
The cleaned dataframe
}
\description{
Read in delimited data
Drop any empty columns
Remove spaces from column names
}
