% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_vipr_ncbi.R
\name{uniqMerge}
\alias{uniqMerge}
\title{Merge unique items in column by group, probably used in a summarize function}
\usage{
uniqMerge(vc, delim = ",")
}
\arguments{
\item{vc}{Vector of values for a particular column}

\item{delim}{Delimiter between unique values, will split original string into delims, and smash them together again}
}
\value{
New string of unique values, delimited by delim
}
\description{
Merge unique items in column by group, probably used in a summarize function
}
