% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/violins.R
\name{violinplot}
\alias{violinplot}
\title{Generate violin plots of a given dataset}
\usage{
violinplot(df, ids = NULL)
}
\arguments{
\item{df}{Dataframe containing HI titer values, each column is a different strain}

\item{ids}{contain the annotation columns, not to be melted}
}
\value{
The ggplot of the violins
}
\description{
Generate violin plots of a given dataset
}
