% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_vipr_ncbi.R
\name{merge_two}
\alias{merge_two}
\title{Merge two dataframes}
\usage{
merge_two(one_df, two_df)
}
\arguments{
\item{one_df}{First data frame to merge, col names will be listed first.}

\item{two_df}{Second data frame to merge, col names will be harmonized with first.}
}
\value{
Merged data frame
}
\description{
Merge two dataframes
}
